QCompare: improve stdlib compatibility checks
Rather than using MS-STL value for Unordered as a "fallback" (not
libstdc++, not libc++), check for it. This means that we now know the
values for the three major stdlib implementations so we can be BC with
them.

This leaves us with toolchains (VxWorks, INTEGRITY) that still don't
have a C++20 stdlib, so we don't know what value to pick. INTEGRITY
will use libc++ for C++20, so just use a matching value. For VxWorks
just pick a value. It's unlikely that it'll cause BC concerns, given
it uses static linking anyhow.

Finally, add an error as a fallback for any other yet unknown stdlib
implementation.

Change-Id: If65e84b4b644a4495943385ee631cf343db8b654
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>